.claude/skills/daily-article-summarizer/pyproject.toml
[project]
name = "daily-article-summarizer"
version = "1.0.0"
description = "Summarize daily articles from mymind directory and push to Notion"
readme = "SKILL.md"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31.0",
    "PyYAML>=6.0",
    "python-dotenv>=1.0.0",
    "notion-client>=2.2.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
